fix(post): update dependency versions (websites using swift)
The version constraints were not strict enough
for the pointfreeco dependencies due to their constant
changing. One issue is that swift-web hasn't been updated
to Swift 5 at this time. See this open issue:
https://github.com/pointfreeco/swift-web/issues/187.

As a result, this project would no longer compile. The solution
is to update the dependencies to be stricter, more compatible, versions
which were found by looking at the pointfreeco main repo's Pakcage.swift:
https://github.com/pointfreeco/pointfreeco/blob/master/Package.swift.

Note: the above link dependencies may change beyond the basis of this fix,
but it is good to note where the source of this fix came from at this point in time.

Example repo also updated with this fix:
https://github.com/jasonzurita/Swift-AWS-Lambda-Website