feat: add resources for workshop
This workshop is part of try! Swift World:
https://www.tryswift.co/world/

Workshop description:
Making a DSL in Swift
Friday, June 5th 1:00pm - 3:00pm EDT
Jason Zurita

Did you know that SwiftUI is a Domain-Specific Language (DSL)!? A DSL is a language that is targeted to a particular kind of problem — like building user interfaces in SwiftUI's case or managing dependencies in CocoaPods (i.e., the Podfile).

In this workshop, we will explore what it takes to create DSLs in Swift. Make sure you have the latest Xcode installed and bring your love for Swift!